+++ src/linter/BaseLinter.ts
@@ -1,21 +1,16 @@
 import { Disposable, workspace, TextDocument, DiagnosticCollection, Diagnostic, languages } from "vscode";
 
 export default abstract class BaseLinter {
-	protected subscriptions: Disposable[];
 	protected diagnostic_collection: DiagnosticCollection;
+	name : string;
 
-	constructor(readonly name: string) {
+	constructor(name: string) {
 		this.diagnostic_collection = languages.createDiagnosticCollection();
-
-		workspace.onDidOpenTextDocument(this.startLint, this, this.subscriptions);
-		workspace.onDidSaveTextDocument(this.startLint, this, this.subscriptions);
-		workspace.onDidCloseTextDocument(this.removeFileDiagnostics, this, this.subscriptions)
+		this.name = name;
 	}
 
 	public startLint(doc: TextDocument) {
-		if (doc.languageId == "verilog") {
-			this.lint(doc);
-		}
+		this.lint(doc);
 	}
 
 	public removeFileDiagnostics(doc: TextDocument) {